rpc.mjs as a library
